Moved string operations to a new text.py file
